requirements.txt
Pillow~=9.2.0
streamlit~=1.16.0